add more unittest
