+++ handler/init_test.go
@@ -1,0 +1,19 @@
+package handler_test
+
+import (
+	"testing"
+
+	"github.com/m-lab/annotation-service/handler"
+)
+
+func TestExtractDateFromFilename(t *testing.T) {
+	date, err := handler.ExtractDateFromFilename("Maxmind/2017/05/08/20170508T080000Z-GeoLiteCity.dat.gz")
+	if date.Year() != 2017 || date.Month() != 5 || date.Day() != 8 || err != nil {
+		t.Errorf("Did not extract data correctly. Expected %d, got %v, %+v.", 20170508, date, err)
+	}
+
+	date2, err := handler.ExtractDateFromFilename("Maxmind/2017/10/05/20171005T033334Z-GeoLite2-City-CSV.zip")
+	if date2.Year() != 2017 || date2.Month() != 10 || date2.Day() != 5 || err != nil {
+		t.Errorf("Did not extract data correctly. Expected %d, got %v, %+v.", 20171005, date2, err)
+	}
+}
